+++ docs/src/modules/components/withRoot.js
@@ -186,10 +186,13 @@ const pages = [
         title: 'About The Lab',
       },
       {
+        pathname: '/lab/slider',
+      },
+      {
         pathname: '/lab/speed-dial',
       },
       {
-        pathname: '/lab/slider',
+        pathname: '/lab/toggle-button',
       },
       findPages[2].children[1],
     ],
