+++ docs/src/pages/lab/toggle-button/toggle-button.md
@@ -1,0 +1,13 @@
+---
+components: ToggleButton, ToggleButtonGroup
+---
+
+## Toggle Buttons
+
+[Toggle buttons](https://material.io/guidelines/components/buttons.html#buttons-toggle-buttons) may be used to group related options. Arrange layout and spacing to convey that certain toggle buttons are part of a group.
+
+The `ToggleButtonGroup` will control the selected of its child buttons when
+given its own `value` prop.
+
+{{"demo": "pages/lab/toggle-button/ToggleButtons.js"}}
+
