+++ packages/material-ui-lab/src/ToggleButton/ToggleButtonGroup.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import withStyles from '../styles/withStyles';
+import withStyles from 'material-ui/styles/withStyles';
 
 export const styles = theme => ({
   root: {
